requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
mlflow>=2.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
scipy>=1.11.0
implicit>=0.7.2
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
jupyterlab>=4.0.0
openpyxl>=3.1.0
